remove MainHttpClient
