+++ BililiveRecorder.WPF/Pages/AdvancedSettingsPage.xaml.cs
@@ -66,31 +66,18 @@ namespace BililiveRecorder.WPF.Pages
 
         private async Task TestCookieAsync()
         {
+            bool succeed;
+            string message;
+
             if (this.httpApiClient is null)
-            {
-                MessageBox.Show("No Http Client Available", "Cookie Test - Failed", MessageBoxButton.OK, MessageBoxImage.Warning);
-                return;
-            }
+                (succeed, message) = (false, "No Http Client Available");
+            else
+                (succeed, message) = await this.httpApiClient.TestCookieAsync().ConfigureAwait(false);
 
-            var resp = await this.httpApiClient.MainHttpClient.GetStringAsync("https://api.live.bilibili.com/xlive/web-ucenter/user/get_user_info").ConfigureAwait(false);
-            var jo = JObject.Parse(resp);
-            if (jo["code"]?.ToObject<int>() != 0)
-            {
-                MessageBox.Show("Response:\n" + resp, "Cookie Test - Failed", MessageBoxButton.OK, MessageBoxImage.Warning);
-                return;
-            }
-
-            var b = new StringBuilder();
-            b.Append("User: ");
-            b.Append(jo["data"]?["uname"]?.ToObject<string>());
-            b.Append("\nUID (from API response): ");
-            b.Append(jo["data"]?["uid"]?.ToObject<string>());
-            b.Append("\nUID (from Cookie): ");
-            b.Append(this.httpApiClient.GetUid());
-            b.Append("\nBUVID3 (from Cookie): ");
-            b.Append(this.httpApiClient.GetBuvid3());
-
-            MessageBox.Show(b.ToString(), "Cookie Test - Successed", MessageBoxButton.OK, MessageBoxImage.Information);
+            if (succeed)
+                MessageBox.Show(message, "Cookie Test - Succeed", MessageBoxButton.OK, MessageBoxImage.Information);
+            else
+                MessageBox.Show(message, "Cookie Test - Failed", MessageBoxButton.OK, MessageBoxImage.Warning);
         }
 
         private void TestScript_Click(object sender, RoutedEventArgs e)
